(portal-ui) Hide saved analysis tab if no analyses are available, add logic to handle analyese deletion, minor tweaks

+++ dcc-portal-ui/src/main/js/app/scripts/analysis/analysisList.js
@@ -20,7 +20,7 @@
 
   var module = angular.module('icgc.analysis.controllers');
 
-  /** 
+  /**
    * Controls list of existing analyses (bench)
    */
   module.controller('AnalysisListController', function($window, $location, AnalysisService) {
@@ -60,15 +60,47 @@
     _this.analysisName = AnalysisService.analysisName;
 
 
-    _this.remove = function(id) {
+    _this.remove = function(id, selectedId) {
       var confirmRemove = window.confirm(REMOVE_ONE);
       if (! confirmRemove) {
         return;
       }
 
-      if (AnalysisService.remove(id) === true) {
-        $location.path('analysis');
+      // 1) Deletion of item that is not currently selected
+      if (id !== selectedId) {
+        AnalysisService.remove(id);
+        return;
       }
+
+
+      // 2) Delete of item we are on, need to find out where to move next
+      var hasNext = _this.analysisList.length > 1? true : false;
+
+      if (hasNext === true) {
+        var currentIndex = _.findIndex(_this.analysisList, function(analysis) {
+          return analysis.id === id;
+        });
+        var max = _this.analysisList.length - 1;
+        var nextIndex = currentIndex <  max ? currentIndex+1 : currentIndex-1;
+
+        console.log('current index', currentIndex);
+        console.log('max', max);
+        console.log('next will be', nextIndex);
+
+        var nextAnalysis = _this.analysisList[nextIndex];
+        if (AnalysisService.remove(id) === true) {
+          $location.path('analysis/view/' + nextAnalysis.type + '/' + nextAnalysis.id);
+        }
+
+      } else {
+        if (AnalysisService.remove(id) === true) {
+          $location.path('analysis');
+        }
+      }
+
+      /*if (AnalysisService.remove(id) === true) {
+        $location.path('analysis');
+      }*/
     };
 
     _this.analysisList = AnalysisService.getAll();
